fix: upgrade rusqlite 0.31 → 0.40 (#465)
Breaking change in rusqlite 0.32+: usize no longer implements
FromSql/ToSql. All COUNT(*) results and LIMIT/OFFSET bindings
must use i64 explicitly.

Changes:
- COUNT(*) queries: row.get::<_, i64>(0) → cast as usize
- LIMIT/OFFSET params: limit as i64, offset as i64
- TagInfo.count, RoomStats.memory_count: i64 intermediate
- count_by_tag: refactored to i64 match → Ok(count as usize)

All 271 tests pass.

+++ crates/uteke-core/src/memory/fts5.rs
@@ -113,7 +113,7 @@ impl super::Store {
                     .prepare(sql)
                     .map_err(|e| Error::db("prepare FTS5 search", e))?;
                 let rows = stmt
-                    .query_map(params![fts_query, ns, limit], |row| {
+                    .query_map(params![fts_query, ns, limit as i64], |row| {
                         let memory = row_to_memory(row)?;
                         let rank: f64 = row.get(14)?;
                         Ok((memory, rank))
@@ -130,7 +130,7 @@ impl super::Store {
                     .prepare(sql)
                     .map_err(|e| Error::db("prepare FTS5 search", e))?;
                 let rows = stmt
-                    .query_map(params![fts_query, limit], |row| {
+                    .query_map(params![fts_query, limit as i64], |row| {
                         let memory = row_to_memory(row)?;
                         let rank: f64 = row.get(14)?;
                         Ok((memory, rank))
@@ -196,7 +196,7 @@ impl super::Store {
                     .prepare(sql)
                     .map_err(|e| Error::db("prepare FTS5 token search", e))?;
                 let rows = stmt
-                    .query_map(params![fts_query, ns, limit], |row| {
+                    .query_map(params![fts_query, ns, limit as i64], |row| {
                         let memory = row_to_memory(row)?;
                         let rank: f64 = row.get(14)?;
                         Ok((memory, rank))
@@ -213,7 +213,7 @@ impl super::Store {
                     .prepare(sql)
                     .map_err(|e| Error::db("prepare FTS5 token search", e))?;
                 let rows = stmt
-                    .query_map(params![fts_query, limit], |row| {
+                    .query_map(params![fts_query, limit as i64], |row| {
                         let memory = row_to_memory(row)?;
                         let rank: f64 = row.get(14)?;
                         Ok((memory, rank))
